Smallinteger field is to small for phone numbers, changed to normal integer

+++ chemie/customprofile/models.py
@@ -58,7 +58,7 @@ class Profile(models.Model):
                                                            default=RELATIONSHIP_STATUS.SINGLE,
                                                            verbose_name="Samlivsstatus")
 
-    phone_number = models.PositiveSmallIntegerField(verbose_name="Mobilnummer")
+    phone_number = models.PositiveIntegerField(verbose_name="Mobilnummer")
     access_card = models.CharField(max_length=10, unique=True, verbose_name="Studentkortnummer")
 
     image_primary = ImageField(upload_to='avatars')
